+++ src/ViKon/Auth/Exception/InvalidKeeperGuardException.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace ViKon\Auth\Exception;
+
+/**
+ * Class InvalidKeeperGuardException
+ *
+ * @package Exception
+ *
+ * @author  Kovács Vince<vincekovacs@hotmail.com>
+ */
+class InvalidKeeperGuardException extends \Exception
+{
+
+}