+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 **Note**: This file is up to date only on master branch.
 
-# Version 3.0-RC
+# Version 3.0.1
+
+- Log when login redirect occurred
+- Add `Keeper` contract
+- Automatically load dependent service providers
+
+# Version 3.0.0
 
 - **Laravel 5.2** support
 - New `PermissionMiddleware` and `LoginRedirectorMiddleware`